pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eval-banana"
version = "0.1.0"
description = "Lightweight aspect-based evaluation framework."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1",
    "httpx>=0.28",
    "openai>=1.78",
    "pydantic>=2.11",
    "PyYAML>=6.0.2",
]

[project.scripts]
eb = "eval_banana.cli:main"
eval-banana = "eval_banana.cli:main"

[project.optional-dependencies]
dev = [
    "pyright>=1.1",
    "pytest>=8.3",
    "ruff>=0.4",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eval_banana"]

[tool.ruff]
src = ["src", "tests"]
target-version = "py312"
line-length = 88

[tool.ruff.format]
skip-magic-trailing-comma = true

[tool.ruff.lint]
extend-select = ["I", "B", "A", "ERA"]

[tool.ruff.lint.isort]
case-sensitive = false
force-single-line = true
force-sort-within-sections = true
known-first-party = ["eval_banana"]
order-by-type = false
relative-imports-order = "closest-to-furthest"
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
extraPaths = ["src"]
reportArgumentType = true
reportMissingParameterType = true
reportCallIssue = true
reportMissingImports = true
reportUnusedImport = true
reportDuplicateImport = true
reportAttributeAccessIssue = true
reportUnnecessaryIsInstance = true
reportUnnecessaryComparison = true
useLibraryCodeForTypes = true
reportGeneralTypeIssues = false
reportOptionalMemberAccess = false
reportOptionalOperand = false
reportOptionalSubscript = false
reportOptionalCall = false
reportUnknownParameterType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownMemberType = false
reportUnknownLambdaType = false
reportIncompatibleMethodOverride = false
reportIncompatibleVariableOverride = false
reportOperatorIssue = false
reportIndexIssue = false
reportReturnType = false
reportAssignmentType = false
reportPrivateUsage = false
reportTypeCommentUsage = false
reportMatchNotExhaustive = false
include = ["src", "tests"]
exclude = ["**/.venv", "**/__pycache__", "tmp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
